+++ kong/rbac_role.go
@@ -3,7 +3,7 @@ package kong
 // RBACRole represents an RBAC Role in Kong.
 // +k8s:deepcopy-gen=true
 type RBACRole struct {
-	CreatedAt *int    `json:"created_at,omitempty" yaml:"created_at,omitempty"`
+	CreatedAt *float64    `json:"created_at,omitempty" yaml:"created_at,omitempty"`
 	ID        *string `json:"id,omitempty" yaml:"id,omitempty"`
 	Name      *string `json:"name,omitempty" yaml:"name,omitempty"`
 	Comment   *string `json:"comment,omitempty" yaml:"comment,omitempty"`
